@@ -23,7 +23,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"            # for structural pattern matching since 3.10
 lxml = "^4.9.2"
 Scrapy = "^2.9.0"
 pymysql = "^1.0.3"
@@ -46,6 +46,8 @@ bs4 = "^0.0.1"
 cchardet = "^2.1.7"
 boto3 = "^1.26.151"
 loguru = "^0.5.3"
+pandas = "^2.0.0"
+numpy = "^1.24.0"
 
 [tool.poetry.dev-dependencies]
 black = "^23.3.0"
@@ -64,5 +66,13 @@ news-please-cc = "newsplease.examples.commoncrawl:main"
 
 [tool.pylint.messages_control]
 max-line-length = 88
-disable = ["missing-docstring", "too-few-public-methods"]
+disable = [
+    "missing-docstring",
+    "too-few-public-methods",
+    "too-many-arguments",
+    "too-many-instance-attributes",
+]
 extension-pkg-whitelist = ["pydantic"]
+
+[tool.pylint.BASIC]
+good-names = ["a", "b", "c", "d", "e", "df", "i", "logger", "n", "tz"]
